Use custom screen_exchange file
The filename will be /tmp/logbuch-screen-exchange-$LOGNAME.  We'll have
to see if LOGNAME is always set or if we need to try more variations.

On most server, LOGNAME will be "root".  On the one hand, it makes sense
to make one user share one screen_exchange file; on the other hand,
maybe we should use the EMAIL variable instead.