Set file permissions as part of copy operation
Before, the screen exchange file was set up (touch & chmod) by the
:LogbuchExchange command.

It makes more sense to do this with the first actual copy operation
because there is no guarantee that :LogbuchExchange will be run before
that.

This patch simply runs touch and chmod with every copy.  We might want
to check the access rights before blindly setting them but for now we'll
keep it simple.