screen exchange: Remove check for known shell
The quick and dirty check for known shells was a crutch, and it turns
out it does more harm than good, so I've removed it again.

I wish I had better documented why I thought we needed it in the first
place.  Right now it seems to me that checking whether we're running in
a screen session -- which we do -- should be good enough.