Remove startup message
The message was added for a use-case in which logbuch.vim was going to
be installed directly on servers where unfamiliar admins might need help
to get started.

The main use-case has since then shifted to remote usage.  This message
is not only unnecessary in this case but gets in the way of important
warnings about existing swap files.