Add tmux interaction
Analogous to the GNU screen exchange feature, this change adds the same
capability for tmux, i.e., a quick way to copy text to a copy/paste
buffer.

Since tmux's buffers are a lot more powerful and easier to use, this
feature is very small and less of a hack.  It might be a good idea to
remove the screen feature or at least make the tmux/screen feature an
either/or option.